# internal/swagger/swagger.yaml
definitions:
...
  github_com_tokenlive_tokenlive-admin_pkg_errors.Error:
    properties:
      code:
        type: integer
      detail:
        type: string
      id:
        type: string
      status:
        type: string
    type: object
...
  schema.OpType:
    enum:
    - EQUAL
    - NOT_EQUAL
    - IN
    - NOT_IN
    - REGULAR
    - PREFIX
    type: string
    x-enum-varnames:
    - OpType_EQUAL
    - OpType_NOT_EQUAL
    - OpType_IN
    - OpType_NOT_IN
    - OpType_REGULAR
    - OpType_PREFIX
...
  schema.PolicyCopyToModelForm:
    properties:
      model_id:
        maxLength: 20
        type: string
      name:
        maxLength: 128
        type: string
      priority:
        minimum: 0
        type: integer
      scope_code:
        maxLength: 128
        type: string
      scope_type:
        maxLength: 32
        type: string
    required:
    - model_id
    type: object
...
  schema.PolicyRoute:
    properties:
      created_at:
        type: string
      creator:
        type: string
      description:
        type: string
      details:
        items:
          $ref: '#/definitions/schema.PolicyRouteDetail'
        type: array
      enabled:
        type: integer
      id:
        type: string
      model_id:
        type: string
      modifier:
        type: string
      name:
        type: string
      priority:
        type: integer
      scope_code:
        type: string
      scope_type:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
  schema.PolicyRouteDetail:
    properties:
      conditions:
        type: string
      created_at:
        type: string
      description:
        type: string
      destinations:
        type: string
      enabled:
        type: integer
      id:
        type: string
      order:
        type: integer
      relation_type:
        type: string
      route_id:
        type: string
      updated_at:
        type: string
    type: object
  schema.PolicyRouteDetailForm:
    properties:
      conditions:
        description: Match conditions
        items:
          $ref: '#/definitions/schema.TagCondition'
        type: array
      created_at:
        description: Create timestamp
        type: string
      description:
        description: Details
        type: string
      destinations:
        description: Destination rules
        items:
          $ref: '#/definitions/schema.TagDestination'
        type: array
      enabled:
        description: Enabled
        type: integer
      id:
        type: string
      order:
        description: Sort order
        type: integer
      relation_type:
        description: Relation type
        maxLength: 20
        type: string
      route_id:
        description: Route ID
        maxLength: 20
        type: string
      updated_at:
        description: Update timestamp
        type: string
    required:
    - relation_type
    - route_id
    type: object
  schema.PolicyRouteForm:
    properties:
      created_at:
        description: Create timestamp
        type: string
      creator:
        description: Creator
        type: string
      description:
        description: Description
        type: string
      details:
        description: RouteDetail
        items:
          $ref: '#/definitions/schema.PolicyRouteDetailForm'
        type: array
      enabled:
        description: Enabled
        type: integer
      id:
        description: Unique ID
        type: string
      model_id:
        description: Owner model ID; empty means template
        maxLength: 20
        type: string
      modifier:
        description: Modifier
        type: string
      name:
        description: Policy name
        maxLength: 128
        type: string
      priority:
        description: Priority
        minimum: 0
        type: integer
      scope_code:
        description: Scope code
        maxLength: 128
        type: string
      scope_type:
        description: Scope type
        maxLength: 32
        type: string
      updated_at:
        description: Update timestamp
        type: string
      version:
        description: Version
        type: integer
    required:
    - name
    type: object
  schema.PolicyTagging:
    properties:
      actions:
        type: string
      conditions:
        type: string
      created_at:
        type: string
      creator:
        type: string
      description:
        type: string
      enabled:
        type: integer
      id:
        type: string
      model_id:
        type: string
      modifier:
        type: string
      name:
        type: string
      priority:
        type: integer
      relation:
        type: string
      scope_code:
        type: string
      scope_type:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
  schema.PolicyTaggingForm:
    properties:
      actions:
        description: Tagging actions
        items:
          $ref: '#/definitions/schema.TaggingAction'
        type: array
      conditions:
        description: Match conditions
        items:
          $ref: '#/definitions/schema.TagCondition'
        type: array
      created_at:
        description: Create timestamp
        type: string
      creator:
        description: Creator
        type: string
      description:
        description: Details
        type: string
      enabled:
        description: Enabled
        enum:
        - 0
        - 1
        type: integer
      id:
        type: string
      model_id:
        description: Owner model ID; empty means template
        maxLength: 20
        type: string
      modifier:
        description: Modifier
        type: string
      name:
        description: Policy name
        maxLength: 128
        type: string
      priority:
        description: Priority
        minimum: 0
        type: integer
      relation:
        description: Relation type
        enum:
        - AND
        - OR
        type: string
      scope_code:
        description: Scope code
        maxLength: 128
        type: string
      scope_type:
        description: Scope type
        maxLength: 32
        type: string
      updated_at:
        description: Update timestamp
        type: string
      version:
        description: Version
        type: integer
    required:
    - name
    - relation
    type: object
...
  schema.RelationType:
    enum:
    - AND
    - OR
    type: string
    x-enum-varnames:
    - RelationType_AND
    - RelationType_OR
...
  schema.TagCondition:
    properties:
      key:
        type: string
      op_type:
        $ref: '#/definitions/schema.OpType'
      type:
        type: string
      values:
        items:
          type: string
        type: array
    type: object
  schema.TagDestination:
    properties:
      conditions:
        items:
          $ref: '#/definitions/schema.TagCondition'
        type: array
      relation_type:
        $ref: '#/definitions/schema.RelationType'
      weight:
        type: integer
    type: object
  schema.TaggingAction:
    properties:
      key:
        description: 标签名/键名
        type: string
      type:
        description: 动作类型/操作范围，如 TAG, REQ_HEADER, RSP_HEADER 等
        type: string
      value:
        description: 标签值，支持变量插值
        type: string
    required:
    - key
    - value
    type: object
...
  util.ResponseResult:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_tokenlive_tokenlive-admin_pkg_errors.Error'
      success:
        type: boolean
      total:
        type: integer
    type: object
info:
  contact: {}
  description: An admin control center for tokelvie.
  title: tokenlive-admin
  version: v1.0.0
paths:
...
  /api/v1/policy/policy-routes:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.PolicyRoute'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Query policy route list
      tags:
      - PolicyRouteAPI
    post:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyRouteForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyRoute'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Create policy route record
      tags:
      - PolicyRouteAPI
  /api/v1/policy/policy-routes/{id}:
    delete:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Delete policy route record by ID
      tags:
      - PolicyRouteAPI
    get:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyRouteForm'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Get policy route record by ID
      tags:
      - PolicyRouteAPI
    put:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyRouteForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Update policy route record by ID
      tags:
      - PolicyRouteAPI
  /api/v1/policy/policy-routes/{id}/copy-to-model:
    post:
      parameters:
      - description: template id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyCopyToModelForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyRoute'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Copy policy route template to model
      tags:
      - PolicyRouteAPI
  /api/v1/policy/policy-taggings:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.PolicyTagging'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Query policy tagging list
      tags:
      - PolicyTaggingAPI
    post:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyTaggingForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyTagging'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Create policy tagging record
      tags:
      - PolicyTaggingAPI
  /api/v1/policy/policy-taggings/{id}:
    delete:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Delete policy tagging record by ID
      tags:
      - PolicyTaggingAPI
    get:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyTaggingForm'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Get policy tagging record by ID
      tags:
      - PolicyTaggingAPI
    put:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyTaggingForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Update policy tagging record by ID
      tags:
      - PolicyTaggingAPI
  /api/v1/policy/policy-taggings/{id}/copy-to-model:
    post:
      parameters:
      - description: template id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyCopyToModelForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyTagging'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Copy policy tagging template to model
      tags:
      - PolicyTaggingAPI
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"